% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_groupings.R
\name{add_groupings}
\alias{add_groupings}
\title{Add PSNU and Age groupsings}
\usage{
add_groupings(df)
}
\arguments{
\item{df}{site tool data frame to apply to}
}
\description{
Add PSNU and Age groupsings
}
