% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggr_group.R
\name{aggr_group}
\alias{aggr_group}
\title{Aggregate by grouping variables}
\usage{
aggr_group(df, indicator, ...)
}
\arguments{
\item{df}{dataframe to create distributions on}

\item{indicator}{variable to aggregate}

\item{...}{grouping variables to sum by (Age & Sex removed from distro)}
}
\description{
Aggregate by grouping variables
}
