% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_distro.R
\name{apply_distro}
\alias{apply_distro}
\title{Apply Distribution to}
\usage{
apply_distro(df, df_distro, ...)
}
\arguments{
\item{df}{dataframe to apply distributions on}

\item{df_distro}{distribution dataframe created with \code{gen_distro()}}

\item{...}{variables to apply the distribution to}
}
\description{
Apply Distribution to
}
