% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_overall_variance.R
\name{check_overall_variance}
\alias{check_overall_variance}
\title{Check the variable at the OU level}
\usage{
check_overall_variance(df_original, df_updated)
}
\arguments{
\item{df_original}{original, unadjusted dataset}

\item{df_updated}{df with updated distribution}
}
\description{
Check the variable at the OU level
}
