% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_distro.R
\name{gen_distro}
\alias{gen_distro}
\title{Generate Variable distribution across grouping vars}
\usage{
gen_distro(df, indicator, ...)
}
\arguments{
\item{df}{dataframe to create distributions on}

\item{indicator}{variable to aggregate}

\item{...}{grouping variables to sum by (Age & Sex removed from distro)}
}
\description{
Generate Variable distribution across grouping vars
}
