% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tab.R
\name{import_tab}
\alias{import_tab}
\title{Import sheet from site tool}
\usage{
import_tab(path, sheetname)
}
\arguments{
\item{path}{full filepath for the site tool}

\item{sheetname}{sheet name in the site tool}
}
\description{
Import sheet from site tool
}
