% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_orig.R
\name{match_orig}
\alias{match_orig}
\title{Match original columns of the Site Tool}
\usage{
match_orig(df)
}
\arguments{
\item{df}{adjusted dataframe}
}
\description{
Match original columns of the Site Tool
}
