% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peds.R
\docType{data}
\name{peds}
\alias{peds}
\title{Full Peds Age Group}
\format{A vector with 4 age groups:
\describe{
\item{peds}{peds age bands}
}}
\usage{
peds
}
\description{
A vector containing peds (<15) age groups.
}
\keyword{datasets}
