% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peds_o01.R
\docType{data}
\name{peds_o01}
\alias{peds_o01}
\title{Peds Age Group Excluding <01}
\format{A vector with 2 age groups:
\describe{
\item{peds}{peds age bands without <01}
}}
\usage{
peds_o01
}
\description{
A vector containing peds (<15) age groups execpt <01.
}
\keyword{datasets}
