% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_excel.R
\name{write_excel}
\alias{write_excel}
\title{Export New Targets to Excel}
\usage{
write_excel(df, wb, tab_name, filepath = "out/")
}
\arguments{
\item{df}{data frame with new targest to export}

\item{wb}{workbook name, must exist prior to creation, eg \code{wb <- openxlsx::createWorkbook()}}

\item{tab_name}{tab to add}

\item{filepath}{filepath, default = "out/"}
}
\description{
Export New Targets to Excel
}
